Refactoriced TransactionPanel
