+++ savewise-client/src/components/Dashboard/Dashboard.tsx
@@ -12,7 +12,6 @@ import { GetCategoriesInput } from '../../services/category-service';
 import { CategoryTypesId } from '../../common/objects/CategoryTypesId';
 import { Transaction } from '../../common/objects/transactions';
 import IconButton from '@material-ui/core/IconButton';
-import AddCircleOutlineIcon from '@material-ui/icons/AddCircleOutline';
 import SettingsOutlinedIcon from '@material-ui/icons/SettingsOutlined';
 import { SnackbarContext } from '../../common/context/SnackbarContext';
 import { constants } from '../../common/objects/constants';
@@ -22,6 +21,7 @@ import MonthsBalanceChart from '../MonthsBalanceChart/MonthsBalanceChart';
 import { MonthInformation } from '../../common/objects/stats';
 import moment from 'moment';
 import TransactionModal from '../TransactionModal/TransactionModal';
+import TransactionPanel from '../TransactionPanel/TransactionPanel';
 
 export default function Dashboard() {
     const {login} = useContext(LoginContext);
@@ -71,7 +71,7 @@ export default function Dashboard() {
                 const toDate = UtilService.lastDayMonthDate(selectedMonthNumber, selectedYear)
                 TransactionService.GetTransactions(userId, fromDate, toDate, 10).then(rsp => {
                     if (rsp.status.success) {
-                        setTransactions( UtilService.sortTransactionByDate(rsp.transactions));
+                        setTransactions(rsp.transactions);
                         setLoading(false);
                     } else {
                         setSnackbarInfo({ severity: "error", message: rsp.status.errorMessage });
@@ -160,32 +160,6 @@ export default function Dashboard() {
             ))
         } else (<Typography variant="h6" component="h2" style={{ display: 'inline', marginLeft: '15px' }}>No categories to show</Typography>);
     };
-
-    const renderTransactions = (transactions: Transaction[]) => {
-        if (transactions.length) {
-            const sortedByDate = transactions.sort((a, b) => Number(b.id) - Number(a.id));
-            return sortedByDate.map(t => {
-                const isIncome = t.category.categoryType?.id === CategoryTypesId.Incomes ? true : false; 
-                const symbol: string = isIncome ? "+" : "-";
-                const color = isIncome ? 'incomeGreenColor' : 'regularColor';
-                return (
-                    <Grid item
-                        container
-                        direction="row"
-                        justify="space-between"
-                        spacing={0}
-                        key={t.id}
-                    >
-                        <Typography variant="h6" component="h2" style={{ display: 'inline', marginLeft: '15px' }}>{t.description}</Typography>
-                        <Typography variant="h6" component="h2" style={{ display: 'inline', marginRight: '15px' }} className={color}>
-                            { symbol } { t.amount } { constants.currency }
-                        </Typography>
-                    </Grid>
-                )
-            }
-            )
-        } else (<Typography variant="h6" component="h2" style={{ display: 'inline', marginLeft: '15px' }}>No transactions to show</Typography>);
-    }
 
     const onSave = (newTransaction: Transaction) => {
         if (transactions.find(t => t.id === newTransaction.id)) {
@@ -352,37 +326,7 @@ export default function Dashboard() {
                 </Grid>
                 <Grid item>
                     <div className={styles.panel}>
-                        <Grid container
-                            direction="column"
-                            alignItems="center"
-                            justify='space-between'
-                            style={{ height: '95%' }}
-                            >
-                                <Typography variant="h4" style={{ marginTop: '5px', marginBottom: '20px' }} component="h2">
-                                    Last transactions
-                                    <IconButton color="primary" aria-label="add transaction" component="span" onClick={() => setOpenModal(true)}>
-                                        <AddCircleOutlineIcon />
-                                    </IconButton>
-                                </Typography>
-                                <Grid container
-                                direction="column"
-                                style={{ height: '70%' }}
-                                spacing={3}
-                                >
-                                    { loading ? 
-                                    (<CircularProgress color="secondary" />) 
-                                    : 
-                                    (<div className={styles.itemsContainer}>
-                                        {renderTransactions(transactions)}
-                                    </div>)    
-                                }
-                                </Grid>
-                                <Link to="/transactions">
-                                    <Button variant="outlined" color="primary">
-                                        Manage
-                                    </Button>
-                                </Link>
-                            </Grid>
+                        <TransactionPanel transactions={transactions} loading={loading} onSave={onSave} ></TransactionPanel>
                     </div>
                 </Grid>
             <TransactionModal conditionToShow={openModal} handleVisibility={() => setOpenModal(!openModal)} handleSave={(t: Transaction) => onSave(t)} />
