+++ savewise-client/src/components/TransactionPanel/TransactionPanel.module.scss
@@ -1,0 +1,7 @@
+.itemsContainer {
+    width: 90%;
+    height: 100%;
+    margin: 0 auto;
+    overflow-x: hidden;
+    overflow-y: auto;
+}
